refactor: move tooltipPosition to lib layer for reusability
